requirements.txt
absl-py==0.9.0
aiocontextvars==0.2.2
albumentations==0.4.5
appdirs==1.4.4
arrow==1.0.3
astor==0.8.1
attrs==19.3.0
backcall==0.2.0
binaryornot==0.4.4
bleach==3.1.5
boto==2.49.0
boto3==1.14.13
botocore==1.17.13
bottle==0.12.19
cairocffi==1.2.0
CairoSVG==2.5.2
certifi==2019.9.11
cffi==1.14.5
cfgv==3.2.0
chardet==3.0.4
click==7.1.2
click-help-colors==0.9
colorama==0.4.4
coloredlogs==15.0
commonmark==0.9.1
conan==1.34.1
contextvars==2.4
cookiecutter==1.7.2
coverage==5.0.3
cssselect==1.1.0
cssselect2==0.4.1
cycler==0.10.0
cymem==2.0.5
Cython==0.29.14
dataclasses==0.8
dataflow==1.2.1
decorator==4.4.2
defusedxml==0.6.0
deprecation==2.0.7
distlib==0.3.1
distro==1.5.0
docutils==0.15.2
easydict==1.9
elasticsearch==7.12.0
en-core-web-sm==2.0.0
entrypoints==0.3
environs==9.3.2
fasteners==0.16
filelock==3.0.12
flake8==3.7.9
Flask==1.1.2
ftfy==5.9
future==0.16.0
glog==0.3.1
graphviz==0.13
grpcio==1.30.0
h11==0.12.0
httpcore==0.12.3
httpx==0.16.1
humanfriendly==9.1
humanize==3.3.0
identify==2.2.2
idna==2.8
imageio==2.6.1
imgaug==0.2.6
immutables==0.15
importlib-metadata==1.7.0
importlib-resources==5.1.2
ipdb==0.13.7
ipykernel==5.1.3
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.17.1
Jinja2==2.10.1
jinja2-time==0.2.0
jmespath==0.10.0
joblib==0.14.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.0.0
jupyter-core==4.6.3
kestrel==1.5.6
kiwisolver==1.1.0
linklink==0.3.1+cuda10.0.torch150.mvapich2.pmi2
loguru==0.5.3
lvis==0.5.1
lxml==4.6.3
Markdown==3.2.2
MarkupSafe==1.1.1
marshmallow==3.11.1
matplotlib==3.1.1
mccabe==0.6.1
mistune==0.8.4
more-itertools==7.2.0
multiprocessing-logging==0.3.1
murmurhash==1.0.5
nart==0.2.4
nbconvert==5.6.1
nbformat==5.0.7
networkx==2.4
ninja==1.10.0.post2
nltk==3.3
node-semver==0.6.1
nodeenv==1.5.0
notebook==6.0.1
numpy==1.17.3
nvidia-dali==0.14.0
object-detection==2.0
onnx==1.6.0
opencv-python==4.1.1.26
packaging==20.4
pandas==0.25.2
pandocfilters==1.4.2
panopticapi==0.1
parso==0.7.0
patch==1.16
patch-ng==1.17.4
pathy==0.4.0
pavi==1.4.17
petrel-oss-sdk===v2.1.0-0-g5ae3dc0-HEAD
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
Pillow-SIMD==6.0.0.post0
pkginfo==1.5.0.1
plotly==4.14.3
pluginbase==1.0.0
portalocker==2.3.0
poyo==0.5.0
pre-commit==2.11.1
preshed==3.0.5
prettytable==0.7.2
prometheus-client==0.8.0
prompt-toolkit==2.0.10
protobuf==3.10.0
ptyprocess==0.6.0
pybind11==2.6.2
pycephs3client==1.0.0
pycocotools==2.0.0
pycocotools-fix==2.0.0.9
pycodestyle==2.5.0
pycparser==2.20
pydantic==1.7.3
pydotz==1.5.1
pyflakes==2.1.1
pygal==2.4.0
Pygments==2.6.1
PyJWT==1.7.1
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
python-dotenv==0.17.0
python-gflags==3.1.2
python-gitlab==2.6.0
python-slugify==4.0.1
pytorch-pretrained-bert==0.6.2
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==19.0.1
qtconsole==4.5.5
QtPy==1.9.0
readme-renderer==24.0
regex==2021.4.4
requests==2.25.1
requests-toolbelt==0.9.1
retrying==1.3.3
rfc3986==1.4.0
rich==9.12.4
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.2
s3transfer==0.3.3
sacremoses==0.0.44
scikit-image==0.16.2
scikit-learn==0.21.3
scipy==1.1.0
seaborn==0.9.0
Send2Trash==1.5.0
six==1.12.0
smart-open==3.0.0
sniffio==1.2.0
spacy==3.0.5
spring-aux===0.5.1.develop.6c0be108
SpringCommonInterface==1.3.0
springvision==1.1.0
tensorboard==2.0.0
tensorboardX==1.8
tensorwatch==0.9.1
terminado==0.8.2
testpath==0.4.4
text-unidecode==1.3
thinc==8.0.2
timm==0.4.5
tinycss==0.4
tinycss2==1.1.0
tokenizers==0.10.2
toml==0.10.2
torch==1.5.0+cu101
torchvision==0.6.1
tornado==6.0.4
tqdm==4.37.0
traitlets==4.3.3
transformers==4.4.2
twine==2.0.0
typer==0.3.2
typing-extensions==3.7.4.3
urllib3==1.25.6
uvloop==0.14.0
virtualenv==20.4.3
wasabi==0.8.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
zipp==3.1.0
